+++ generator/lib/models/source_assignment.dart
